강원대 BE_고명준 3주차 (2,3단계) (#117)
* refactor(entity): ManyToOne 연관관계 추가 및 Getter 변경

* refactor(entity):  Getter 변경

* refactor: Wishlistcontroller, WishService 로직 수정

* refactor: ProductService 로직 수정

* feat(memberService): login 실패시 exception handle

* refactor: entity클래스 유효성 검사 추가

* docs: README파일 수정

* refactor(service): deleteWishlist 메서드 exception handling

* style: convention 수정

* test:WishRepositoryTest 수정

* test: containsExactlyInAnyOrder 메서드로 수정

* test: productRepositoryTest 수정

* feat(dto): ProductDTO 생성

* feat(controller): 인자를 DTO로 변경

* feat(dto): MemverDTO 생성

* feat(dto): WishDTO 생성

* feat(dto): WishDTO toEntity 생성

* refactor(controller): WishController 인자 WishDTO로 변경

* feat(ProductRepository): Page반환 메서드 추가

* feat(ProductService): Page반환 메서드 추가

* refactor(ProductService): Page반환 메서드 수정

* feat(Controller): Product Page반환 컨트롤러 추가

* feat(WishRepository): Page반환 메서드 추가

* feat(WishService): Page반환 서비스 메서드 추가

* refactor(controller): ProductController 반환타입 ResponseEntity로 변경

* feat(WishController): Page 반환 Controller메서드 작성

* refactor(WishlistController): 잘못된 임포트 수정

* Update README.md

+++ src/main/java/gift/service/WishlistService.java
@@ -1,8 +1,19 @@
 package gift.service;
 
+import gift.entity.Member;
 import gift.entity.Wish;
+import gift.exception.DataNotFoundException;
+import gift.repository.MemberRepository;
 import gift.repository.WishRepository;
+
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,23 +22,50 @@ import org.springframework.transaction.annotation.Transactional;
 public class WishlistService {
 
     private final WishRepository wishRepository;
+    private final MemberRepository memberRepository;
+    private final int PAGE_SIZE = 5;
 
-    public WishlistService(WishRepository wishRepository){
-        this.wishRepository=wishRepository;
+    public WishlistService(WishRepository wishRepository, MemberRepository memberRepository) {
+        this.wishRepository = wishRepository;
+        this.memberRepository = memberRepository;
     }
 
-    public void addWishlist(Wish wish) {
-        wishRepository.save(wish);
+    public void addWishlist(Wish wish, String email) {
+        Member wishMember = wish.getMember();
+        Member authMember = memberRepository.findByEmail(email);
+        if (wishMember.getId().equals(authMember.getId())) {
+            wishRepository.save(wish);
+        }
 
     }
 
-    public List<Wish> getWishlist(Long id) {
-        return wishRepository.findByMemberId(id);
+
+    public void deleteWishlist(Long id, String email) {
+        Member authMember = memberRepository.findByEmail(email);
+        Optional<Wish> wish = wishRepository.findById(id);
+        if (wish.isEmpty()) {
+            throw new DataNotFoundException("존재하지 않는 Wishlist: wishList를 삭제할 수 없습니다.");
+        }
+        Wish deletingWish = wish.get();
+
+        //삭제하려는 Wish의 MemberId와 accessToken에서 받은 memberId가 같아야 삭제
+        if (authMember.getId().equals(deletingWish.getMember().getId())) {
+            wishRepository.deleteById(id);
+        }
+
     }
 
-    public void deleteWishlist(Long id){
-         wishRepository.deleteById(id);
+    public List<Wish> getWishlistByEmail(String email) {
+        Member member = memberRepository.findByEmail(email);
+        return wishRepository.findByMemberId(member.getId());
     }
 
+    public Page<Wish> getWishPage(String email, int page){
+        List<Sort.Order> sorts = new ArrayList<>();
+        sorts.add(Sort.Order.asc("id"));
+        Pageable pageable = PageRequest.of(page,PAGE_SIZE,Sort.by(sorts));
+        Member member = memberRepository.findByEmail(email);
+        return wishRepository.findByMemberId(member.getId(),pageable);
 
+    }
 }
